refactor: remove unused package
