+++ main.go
@@ -1,5 +1,5 @@
 package main
 
-import kcl_openapi "kusionstack.io/kcl-openapi/pkg/cmds/kcl-openapi"
+import "kusionstack.io/kcl-openapi/pkg/cmds"
 
-func main() { kcl_openapi.Main() }
+func main() { cmds.Main() }
